+++ object.go
@@ -6,10 +6,10 @@ import (
 
 type Object struct {
 	// X is the logical X coordinate of the object in the current view.
-	X int
+	X float64
 
 	// Y is the logical Y coordinate of the object in the current view.
-	Y int
+	Y float64
 
 	// Key is the object's identifier.
 	Key string
@@ -65,7 +65,7 @@ func createSingleObject(x, y float64, groundBound bool) singleObject {
 
 func (so *singleObject) ToObjects(cam Camera) []Object {
 	pos := cam.ScreenPosition(coords.VectorFromCartesian(so.sox, so.soy))
-	so.cache[0].X, so.cache[0].Y = int(pos.X()), int(pos.Y())
+	so.cache[0].X, so.cache[0].Y = pos.X(), pos.Y()
 	return so.cache
 }
